+++ src/ga4gh/gks/metaschema/tools/release_prep/git.py
@@ -1,9 +1,8 @@
-"""Git and submodule helpers for GKS release preparation.
+"""Git submodule metadata and checkout helpers for release preparation.
 
-This module owns the git-specific mechanics used by ``gks-release-prep``:
-parsing ``.gitmodules``, resolving the single immediate upstream submodule,
-checking clean working trees, updating submodules from remote branches, and
-choosing the highest reachable semantic-version tag.
+This module parses and updates ``.gitmodules``, resolves the single immediate
+upstream submodule, updates it from its remote branch, and chooses the highest
+reachable semantic-version tag. Worktree-status checks live in ``worktree``.
 """
 
 import re
@@ -14,18 +13,20 @@ from typing import Callable
 
 from packaging.version import InvalidVersion, Version
 
+from ga4gh.gks.metaschema.tools.release_prep.worktree import (
+    CommandOutputRunner,
+    Reporter,
+    require_clean_worktree,
+    warn_if_worktree_dirty,
+)
+
 CommandRunner = Callable[[list[str], Path], None]
-CommandOutputRunner = Callable[[list[str], Path], str]
-Reporter = Callable[[str], None]
 
 GITMODULES_FN = ".gitmodules"
 ORIGIN_REMOTE = "origin"
 SCHEMA_DIR_NAME = "schema"
 SUBMODULES_DIR_NAME = "submodules"
-GIT_STATUS_PORCELAIN_COMMAND = ("git", "status", "--porcelain")
 GIT_FETCH_ALL_TAGS_COMMAND = ("git", "fetch", "--all", "--tags")
-GIT_UPSTREAM_BRANCH_COMMAND = ("git", "rev-parse", "--abbrev-ref", "--symbolic-full-name", "@{u}")
-GIT_UPSTREAM_COUNTS_COMMAND = ("git", "rev-list", "--left-right", "--count", "HEAD...@{u}")
 
 SEMVER_TAG_RE = re.compile(
     r"^v?(?P<major>0|[1-9]\d*)\."
@@ -351,89 +352,6 @@ def update_gitmodules_branch(entry: GitmodulesEntry, branch: str) -> None:
         lines[entry.branch_line] = branch_line
 
     entry.gitmodules_fp.write_text("\n".join(lines) + "\n", encoding="utf-8")
-
-
-def warn_if_worktree_dirty(
-    repo_dir: Path,
-    label: str,
-    output_runner: CommandOutputRunner,
-    reporter: Reporter | None,
-) -> None:
-    """Warn if a git working tree has uncommitted changes.
-
-    :param repo_dir: Git repository directory to inspect.
-    :param label: Human-readable repository label for error messages.
-    :param output_runner: Command runner used for git output.
-    :param reporter: Optional progress reporter.
-    :raises subprocess.CalledProcessError: If git status fails.
-    """
-    status = output_runner(list(GIT_STATUS_PORCELAIN_COMMAND), repo_dir)
-
-    if not status or reporter is None:
-        return
-
-    reporter(f"Warning: {label} has uncommitted changes. Review the final diff carefully.")
-
-
-def require_clean_worktree(repo_dir: Path, label: str, output_runner: CommandOutputRunner) -> None:
-    """Require a git working tree to have no uncommitted changes.
-
-    :param repo_dir: Git repository directory to inspect.
-    :param label: Human-readable repository label for error messages.
-    :param output_runner: Command runner used for git output.
-    :raises ValueError: If the repository has uncommitted changes.
-    :raises subprocess.CalledProcessError: If git status fails.
-    """
-    status = output_runner(list(GIT_STATUS_PORCELAIN_COMMAND), repo_dir)
-
-    if not status:
-        return
-
-    msg = f"{label} has uncommitted changes. Commit, stash, or discard them before running release prep."
-    raise ValueError(msg)
-
-
-def warn_if_product_branch_not_current(
-    repo_dir: Path, output_runner: CommandOutputRunner, reporter: Reporter | None
-) -> None:
-    """Warn when the product branch may not be current with its upstream.
-
-    This is intentionally non-fatal because release prep should not fetch or
-    merge the product repository.
-
-    :param repo_dir: Product repository root directory.
-    :param output_runner: Command runner used for git output.
-    :param reporter: Optional progress reporter.
-    """
-    try:
-        output_runner(list(GIT_UPSTREAM_BRANCH_COMMAND), repo_dir)
-        counts = output_runner(list(GIT_UPSTREAM_COUNTS_COMMAND), repo_dir)
-    except subprocess.CalledProcessError:
-        if reporter is not None:
-            reporter("Warning: product branch has no upstream tracking branch; unable to check if it is current.")
-        return
-
-    ahead_text, _separator, behind_text = counts.partition("\t")
-    if not behind_text:
-        ahead_text, _separator, behind_text = counts.partition(" ")
-
-    try:
-        ahead = int(ahead_text)
-        behind = int(behind_text)
-    except ValueError:
-        if reporter is not None:
-            reporter(f"Warning: unable to parse product branch upstream status: {counts}")
-        return
-
-    if reporter is None:
-        return
-
-    if behind and ahead:
-        reporter(f"Warning: product branch has diverged from upstream ({ahead} ahead, {behind} behind).")
-    elif behind:
-        reporter(f"Warning: product branch is {behind} commit(s) behind upstream.")
-    elif ahead:
-        reporter(f"Warning: product branch is {ahead} commit(s) ahead of upstream.")
 
 
 def require_submodule_dir(entry: GitmodulesEntry) -> Path:
